게시판 crud 파일업로드/다운로드 crud

+++ src/main/java/knou/seoul/hanwoori/domain/post/PostController.java
@@ -1,6 +1,9 @@
 package knou.seoul.hanwoori.domain.post;
 
 import jakarta.servlet.http.HttpSession;
+import knou.seoul.hanwoori.domain.file.FileService;
+import knou.seoul.hanwoori.domain.file.dto.File;
+import knou.seoul.hanwoori.domain.file.dto.FileSource;
 import knou.seoul.hanwoori.domain.member.dto.Member;
 import knou.seoul.hanwoori.domain.post.dto.Post;
 import knou.seoul.hanwoori.domain.signup.dto.Signup;
@@ -13,7 +16,10 @@ import org.springframework.ui.Model;
 import org.springframework.validation.BindingResult;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.multipart.MultipartFile;
 
+import java.io.IOException;
+import java.nio.file.AccessDeniedException;
 import java.util.List;
 import java.util.Optional;
 
@@ -25,6 +31,7 @@ import static knou.seoul.hanwoori.common.SessionConst.LOGIN_MEMBER;
 public class PostController {
 
     private final PostService postService;
+    private final FileService fileService;
 
     @ModelAttribute("categorys")
     public Post.Category[] categorys() {
@@ -38,7 +45,8 @@ public class PostController {
     }
 
     @PostMapping("/new")
-    public String create(@Validated @ModelAttribute Post post, BindingResult bindingResult, HttpSession session) {
+    public String create(@Validated @ModelAttribute Post post, BindingResult bindingResult, HttpSession session,
+                         @RequestParam MultipartFile[] files) throws IOException {
 
         if (bindingResult.hasErrors()) {
             return "domain/post/post-form";
@@ -49,32 +57,46 @@ public class PostController {
         postService.save(post);
         Long postId = post.getPostId();
 
+        //파일업로드
+        fileService.storeFiles(files, post);
+
         return "redirect:/posts/" + postId;
     }
 
     @GetMapping
-    public String list(Model model) {
-        List<Post> posts = postService.findAll();
+    public String list(@RequestParam(value = "category", defaultValue = "notice") Post.Category category, Model model) {
+        List<Post> posts = postService.findByCategory(category);
         model.addAttribute("posts", posts);
+        model.addAttribute("selectedCategory", category);
         return "domain/post/post-list";
     }
 
     @GetMapping("/{id}")
     public String view(@PathVariable Long id, Model model) {
         Optional<Post> post = postService.findById(id);
+        List<File> files = fileService.findBySourceKindAndId(File.SourceKind.post, id);
+
         model.addAttribute("post", post.orElseGet(Post::new));
+        model.addAttribute("files", files);
         return "domain/post/post-view";
     }
 
     @GetMapping("/{id}/edit")
-    public String editForm(@PathVariable Long id, Model model) {
+    public String editForm(@PathVariable Long id, Model model, HttpSession session) throws AccessDeniedException {
+
+        checkAccessPossible(id,session);
+
         Optional<Post> post = postService.findById(id);
+        List<File> files = fileService.findBySourceKindAndId(File.SourceKind.post, id);
+
         model.addAttribute("post", post.orElseGet(Post::new));
+        model.addAttribute("files", files);
         return "domain/post/post-edit-form";
     }
 
     @PostMapping("/{id}/edit")
-    public String modify(@PathVariable Long id, @Validated @ModelAttribute Post post,BindingResult bindingResult, HttpSession session) {
+    public String modify(@PathVariable Long id, @Validated @ModelAttribute Post post,BindingResult bindingResult, HttpSession session,
+                         @RequestParam MultipartFile[] files) throws IOException {
 
         if(bindingResult.hasErrors()) {
             return "domain/post/post-edit-form";
@@ -82,17 +104,37 @@ public class PostController {
         Optional<Member> loginMember = Optional.ofNullable((Member) session.getAttribute(LOGIN_MEMBER));
         post.setMember(loginMember.orElseGet(Member::new));
         postService.modify(post);
+
+        //파일업로드
+        fileService.storeFiles(files, post);
+
         return "redirect:/posts/" + id;
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<Void> delete(@PathVariable Long id) {
+    public ResponseEntity<Void> delete(@PathVariable Long id, HttpSession session) throws AccessDeniedException {
+
+        checkAccessPossible(id,session);
+
         int count = postService.delete(id);
         if (count == 1) {
             return ResponseEntity.noContent().build();
         } else {
             return ResponseEntity.status(HttpStatus.CONFLICT).build();
         }
+    }
+
+    private void checkAccessPossible(Long id, HttpSession session) throws AccessDeniedException {
+        Optional<Member> loginMember = Optional.ofNullable((Member) session.getAttribute(LOGIN_MEMBER));
+        Member member = loginMember.orElseGet(Member::new);
+
+        Optional<Post> post = postService.findById(id);
+        Post findPost = post.orElseGet(Post::new);
+
+        if(!findPost.getMember().getMemberId().equals(member.getMemberId())){
+            throw new AccessDeniedException("수정/삭제권한이 없습니다.");
+        }
+
     }
 
 }
